pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "project-chimera"
version = "0.1.0"
description = "A Novel Benchmark for Evaluating Strategic Deception and Subtle Sabotage in Multi-Agent LLM Systems"
authors = ["Research Team <research@example.com>"]
readme = "README.md"
packages = [{include = "chimera", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
# Multi-agent frameworks
mesa = "^2.1.0"
agentpy = "^0.1.5"
langgraph = "^0.0.40"
# LLM integration
openai = "^1.0.0"
anthropic = "^0.18.0"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"
# Data and analysis
pandas = "^2.0.0"
numpy = "^1.24.0"
scipy = "^1.10.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.15.0"
# Configuration and utilities
pydantic = "^2.0.0"
pyyaml = "^6.0"
click = "^8.1.0"
rich = "^13.4.0"
python-dotenv = "^1.0.0"
# Logging and monitoring
structlog = "^23.1.0"
opentelemetry-api = "^1.18.0"
opentelemetry-sdk = "^1.18.0"
# Testing and quality
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
pre-commit = "^3.3.0"
sphinx = "^7.1.0"
sphinx-rtd-theme = "^1.3.0"

[tool.poetry.group.interpretability.dependencies]
torch = "^2.0.0"
transformers = "^4.30.0"
datasets = "^2.14.0"
scikit-learn = "^1.3.0"
umap-learn = "^0.5.3"

[tool.poetry.scripts]
chimera = "chimera.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
